bug: fix an issue where mime types are not preserved
